Move Home-new.ts under folder "model"
Home-new class is a typical example of "model" classes. They are often
defined for a type of objects, putting them in a folder should improve
project structure. These classes are also easier to find.
